moved RqlParserException to ditto-model-base as too many other modules would have to add "ditto-model-rql" as dependency otherwise
* split up ditto-model-rql and ditto-model-rqlparser in order to separate btwn model and parser
