+++ model/pom.xml
@@ -37,6 +37,7 @@
         <module>thingsearch-parser</module>
         <module>query</module>
         <module>rql</module>
+        <module>rql-parser</module>
     </modules>
 
     <dependencies>
