+++ model/query/pom.xml
@@ -32,6 +32,10 @@
             <groupId>org.eclipse.ditto</groupId>
             <artifactId>ditto-model-rql</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.eclipse.ditto</groupId>
+            <artifactId>ditto-model-rql-parser</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
