+++ model/rql-parser/pom.xml
@@ -21,15 +21,20 @@
         <version>${revision}</version>
     </parent>
 
-    <artifactId>ditto-model-rql</artifactId>
+    <artifactId>ditto-model-rql-parser</artifactId>
     <packaging>bundle</packaging>
-    <name>Eclipse Ditto :: Model :: RQL</name>
+    <name>Eclipse Ditto :: Model :: RQL parser</name>
 
     <dependencies>
         <!-- ### Compile ### -->
         <dependency>
             <groupId>org.eclipse.ditto</groupId>
-            <artifactId>ditto-model-base</artifactId>
+            <artifactId>ditto-model-rql</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.parboiled</groupId>
+            <artifactId>parboiled_${scala.version}</artifactId>
         </dependency>
 
         <!-- ### Testing ### -->
@@ -44,6 +49,40 @@
     <build>
         <plugins>
             <plugin>
+                <artifactId>maven-compiler-plugin</artifactId>
+            </plugin>
+
+            <plugin>
+                <groupId>net.alchim31.maven</groupId>
+                <artifactId>scala-maven-plugin</artifactId>
+                <configuration>
+                    <scalaVersion>${scala.full.version}</scalaVersion>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>compile</id>
+                        <goals>
+                            <goal>compile</goal>
+                        </goals>
+                        <phase>compile</phase>
+                    </execution>
+                    <execution>
+                        <id>test-compile</id>
+                        <goals>
+                            <goal>testCompile</goal>
+                        </goals>
+                        <phase>test-compile</phase>
+                    </execution>
+                    <execution>
+                        <phase>process-resources</phase>
+                        <goals>
+                            <goal>compile</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+
+            <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
                 <extensions>true</extensions>
@@ -54,7 +93,8 @@
                             org.eclipse.ditto.*
                         </Import-Package>
                         <Export-Package>
-                            org.eclipse.ditto.model.rql.*
+                            !org.eclipse.ditto.model.rqlparser.internal,
+                            org.eclipse.ditto.model.rqlparser.*
                         </Export-Package>
                     </instructions>
                 </configuration>
