+++ model/thingsearch-parser/src/main/java/org/eclipse/ditto/model/thingsearchparser/OptionParser.java
@@ -9,7 +9,7 @@
  * Contributors:
  *    Bosch Software Innovations GmbH - initial contribution
  */
-package org.eclipse.ditto.model.thingsearchparser.options.rql;
+package org.eclipse.ditto.model.thingsearchparser;
 
 import java.util.List;
